Streamlined lint_command() and moved to separate module
